adk_agent/catalog_admin/agent_engine_requirements.txt
google-adk==1.3.0
google-cloud-aiplatform==1.97.0
google-cloud-storage==2.19.0
google-cloud-logging==3.11.4
requests==2.32.4
python-dotenv==1.1.0
pydantic==2.11.7
google-genai==1.20.0
cloudpickle==3.1.1
google-generativeai==0.3.2

